refactored validation schema
